Only turnon if Watt reading is reasonable
One of the devices I use seems to reset, and then show a extremely big unreasonable value as usage.
I have now added that the reading must be below 16A 240V = 3840 Watt (The max a EU socket can provide). and this check is only for the first reading before turning on, so it is very reasonable to make this check.

I also added the reading value to the log